+++ trDynaPS_mb.cpp
@@ -1,3 +1,4 @@
+#include "Variables.hpp"
 #include "globals.hpp"
 #include <nvector/nvector_serial.h>    /* access to serial N_Vector       */
 #include "trDynaPS.hpp"
@@ -32,77 +33,35 @@ int trDynaPS::trDynaPS_mb(realtype t, N_Vector u, N_Vector u_dot, void *user_dat
     realtype *x = N_VGetArrayPointer(u);
     realtype *dxdt = N_VGetArrayPointer(u_dot);
 
-    //global trDynaPS2RedReg_cal
     myVars->trDynaPS2RedReg_cal = 0;
-    
-    //fidw = fopen('Mymatrix2.txt','a');
-    //fprintf(fidw, '//g   ', t);
-    //fprintf(fidw, '//g  ', trDynaPS_Con);
-    //fprintf(fidw, '\n');
-    //fclose(fidw);
-    //N_Vector dy, ddxdt;
-    //dy = N_VNew_Serial(96);
-    //ddxdt = N_VNew_Serial(96);
-    //arr DynaPS_Con = zeros(96);
-    
-    //for (int m = 0; m < 96; m++)
-    //    DynaPS_Con[m] = x[m];
-        //NV_Ith_S(dy, m) = x[m];
+
     trDynaPSCon trDynaPS_con(x);
-    //DynaPSCon DynaPS_con = trDynaPS_con.DynaPS_con;//(x);
-    
-    //arr RROEA_Con = zeros(10);
-    //for (int m = 0; m < 10; m++)
-    //    RROEA_Con[m] = x[m + 110];
-    //RROEACon RROEA_con(RROEA_Con);
-    //sRROEACon RROEA_con = trDynaPS_con.RROEA_con;
-    
+
     Condition(t, myVars);
     const double light = 1.0;
     myVars->FI_Param[0] = light;
     myVars->BF_Param[0] = light;
-    
+
     myVars->RROEA_Param[1] = 1;
     arr RROEA_DYDT = RROEA_Mb(t, trDynaPS_con.RROEA_con, myVars);
-    //fprintf(" //g    ", DynaPS_Con);
-    //fprintf("\n");
-    //DynaPS dps = DynaPS(myVars);
-    
+
     arr DynaPS_DYDT = DynaPSmb(t, trDynaPS_con.DynaPS_con, myVars);
-    
-    //realtype *DynaPS_DYDT = N_VGetArrayPointer(ddxdt);
-    //arr trDynaPS_DYDT = zeros(119);
-    
-    for (int index = 0; index < 96; index++)
+
+    for (size_t index = 0; index < 96; index++)
         dxdt[index] = DynaPS_DYDT[index];
-    
-    
-    for (int index = 0; index < 10; index++)
+
+
+    for (size_t index = 0; index < 10; index++)
         dxdt[index + 110] = RROEA_DYDT[index];
-    
-    
-    //global RROEA2trDynaPS_ve2Fd;
-    //global BF2RROEA_Vbf16;
-    //global RROEA2trDynaPS_veFd2Calvin;
-    //global BF2trDynaPS_vbfn2;
-    
-    //global BF2TrDynaPSMB_vcet;
-    //global AVR;
-    //global PRGlu;
+
+
     //////WY201804
     const double Temp = RROEA_DYDT[8] - myVars->RROEA2trDynaPS_ve2Fd + myVars->BF2RROEA_Vbf16 / myVars->AVR + myVars->RROEA2trDynaPS_veFd2Calvin - myVars->BF2trDynaPS_vbfn2 - myVars->BF2TrDynaPSMB_vcet / myVars->AVR;
-    //Temp = RROEA_DYDT(9) - RROEA2trDynaPS_ve2Fd + BF2RROEA_Vbf16/AVR + RROEA2trDynaPS_veFd2Calvin - BF2trDynaPS_vbfn2 - BF2TrDynaPSMB_vcet/AVR-PRGlu/AVR;
-    
+
     dxdt[118] = Temp * myVars->AVR;
     dxdt[23] = Temp * myVars->AVR;
-    //fid = fopen('Mymatrix.txt','a');
-    //fprintf(fid, '//e   ', t);
-    //fprintf(fid, '//f  ', trDynaPS_DYDT);
-    //fprintf(fid, '\n');
-    //fclose(fid);
-    
+
     GenOut(t, myVars);
-    //N_VDestroy(dy);
-    //N_VDestroy(ddxdt);
+
     return 0;
 }
